+++ CHANGELOG.md
@@ -10,7 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * datamine-types: Add definitions for the following types
 	* `IEvolutionMaterial`
 	* `IMission`, `ClearBonus`, and `IMimicInfo` (and their dependencies)
-	* `IUnit.evolution`, `IUnit.first_clear_missions`, `IUnit.guild_raid`
+	* `IUnitBond`
+	* `IUnit.evolution`, `IUnit.first_clear_missions`, `IUnit.guild_raid`, `IUnit.bonds`
 	* `IBfmtMetadata` (and applied as an optional property on `IUnit`, `IItem`, `ILeaderSkill`, `IBraveBurst`, `IExtraSkill`, and `IMission`)
 
 ## [0.4.1] - 2020-04-12