+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased] - YYYY-MM-DD
+### ⚠ Breaking Changes
+* datamine-types: stat entries that are not `_lord` or `_base` in `IUnit.stats` are now optional.
+
 ### ➕ Additions
 * sp-enhancements: Add type definition to `addedEntries` argument for `getAllDependenciesForSpEntry` and `getAllEntriesThatDependOnSpEntry`
 * datamine-types: Add definitions for the following types
@@ -13,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 	* `IUnitBond`
 	* `IUnit.evolution`, `IUnit.first_clear_missions`, `IUnit.guild_raid`, `IUnit.bonds`
 	* `IBfmtMetadata` (and applied as an optional property on `IUnit`, `IItem`, `ILeaderSkill`, `IBraveBurst`, `IExtraSkill`, and `IMission`)
+	* `UnitType`
 
 ## [0.4.1] - 2020-04-12
 ### ➕ Additions