+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 	* `IUnit.guild_raid`
 	* `IBfmtMetadata` (and applied as an optional property on `IUnit`, `IItem`, `ILeaderSkill`, `IBraveBurst`, `IExtraSkill`, and `IMission`)
 	* `UnitType`
+	* `IAttackInfo`, `TargetAreaShorthand`, and `IBraveBurst.attackInfo`
 
 ## [0.4.1] - 2020-04-12
 ### ➕ Additions