+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased] - YYYY-MM-DD
+
+## [0.5.0] - 2020-05-03
 ### ⚠ Breaking Changes
 * datamine-types: stat entries that are not `_lord` or `_base` in `IUnit.stats` are now optional.
 
@@ -184,7 +186,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 	* `UnitKind` enum
 	* `IUnit` interface
 
-[Unreleased]: https://github.com/BluuArc/bfmt-utilities/compare/v0.4.1...HEAD
+[Unreleased]: https://github.com/BluuArc/bfmt-utilities/compare/v0.5.0...HEAD
+[0.5.0]: https://github.com/BluuArc/bfmt-utilities/compare/v0.4.1...v0.5.0
 [0.4.1]: https://github.com/BluuArc/bfmt-utilities/compare/v0.3.1...v0.4.1
 [0.4.0]: https://github.com/BluuArc/bfmt-utilities/compare/v0.3.1...v0.4.0
 [0.3.1]: https://github.com/BluuArc/bfmt-utilities/compare/v0.3.0...v0.3.1